* Requested to upload INFDEV ON 2.12.22
HQRipper INFDEV is a software designed to rip information from files and folders in order to create an index of the data for easy access. The program can be used to quickly gather information on specific files or folders, or to collect data from an entire drive.

The program operates by scanning through all files and folders within a designated source, extracting any identifying information it can find such as file names, sizes, dates modified, and contents. This data is then compiled into an easily-accessible database for further analysis.

HQRipper INFDEV also includes a number of features specifically designed for Information Security professionals. These features include the ability to extract hashes of files and folders as well as metadata such as author, title, and date created/modified. Additionally, the program can be configured to scan for specific file types or keywords in order to quickly gather targeted data sets.